# apps/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: GoList Sharing API
  version: v1
servers:
  - url: https://api.example.com
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    ShareToken:
      name: shareToken
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ListId:
      name: listId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UpdatedAfter:
      name: updatedAfter
      in: query
      required: true
      schema:
        type: string
        format: date-time
    DeviceIdHeader:
      name: X-Device-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ListItem:
      type: object
      required: [id, name, category, deleted, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        quantityOrUnit:
          type: string
        category:
          type: string
        deleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListDocument:
      type: object
      required: [listId, name, items, createdAt, updatedAt]
      properties:
        listId:
          type: string
          format: uuid
        name:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListCreateRequest:
      type: object
      required: [listId, name]
      properties:
        listId:
          type: string
          format: uuid
        name:
          type: string
    ListCreateResponse:
      type: object
      required: [listId, shareToken]
      properties:
        listId:
          type: string
          format: uuid
        shareToken:
          type: string
          format: uuid
    ItemUpsertRequest:
      type: object
      required: [name, category, deleted, updatedAt]
      properties:
        name:
          type: string
        quantityOrUnit:
          type: string
        category:
          type: string
        deleted:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    UpdatedItemsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListItem"
    ListNameUpdateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ShareTokenEntry:
      type: object
      required: [tokenId, listId, createdAt]
      properties:
        tokenId:
          type: string
          format: uuid
        listId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ShareTokenCreateResponse:
      allOf:
        - $ref: "#/components/schemas/ShareTokenEntry"
        - type: object
          required: [shareToken]
          properties:
            shareToken:
              type: string
              format: uuid
    ShareTokenRedeemResponse:
      type: object
      required: [listId]
      properties:
        listId:
          type: string
          format: uuid
security:
  - BearerAuth: []
paths:
  /v1/lists:
    put:
      summary: Create list (idempotent) or update existing list when caller has access
      security: []
      parameters:
        - $ref: "#/components/parameters/DeviceIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListCreateRequest"
      responses:
        "201":
          description: List created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCreateResponse"
        "200":
          description: Existing list updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCreateResponse"
        "403":
          description: Existing list update requires caller access to that list
  /v1/lists/{listId}:
    parameters:
      - $ref: "#/components/parameters/ListId"
      - $ref: "#/components/parameters/DeviceIdHeader"
    get:
      summary: Fetch list by list id
      responses:
        "200":
          description: List document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDocument"
    delete:
      summary: Delete list (creator device only)
      responses:
        "204":
          description: List deleted
        "403":
          description: Only the creator device may delete the list
  /v1/lists/{shareToken}/name:
    parameters:
      - $ref: "#/components/parameters/ShareToken"
      - $ref: "#/components/parameters/DeviceIdHeader"
    patch:
      summary: Update list name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListNameUpdateRequest"
      responses:
        "204":
          description: List name updated
  /v1/lists/{shareToken}/items:
    parameters:
      - $ref: "#/components/parameters/ShareToken"
      - $ref: "#/components/parameters/DeviceIdHeader"
    get:
      summary: Fetch items updated after timestamp
      parameters:
        - $ref: "#/components/parameters/UpdatedAfter"
      responses:
        "200":
          description: Updated items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedItemsResponse"
  /v1/lists/{shareToken}/items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/ShareToken"
      - $ref: "#/components/parameters/DeviceIdHeader"
      - $ref: "#/components/parameters/ItemId"
    get:
      summary: Fetch item by id
      responses:
        "200":
          description: Item document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListItem"
    put:
      summary: Upsert item in list (create or update by client-provided itemId)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpsertRequest"
      responses:
        "201":
          description: Item created
        "204":
          description: Item updated
        "409":
          description: Item id belongs to another list
  /v1/share-tokens/{shareToken}/redeem:
    parameters:
      - $ref: "#/components/parameters/ShareToken"
      - $ref: "#/components/parameters/DeviceIdHeader"
    post:
      summary: Redeem a share token for the requesting device
      responses:
        "200":
          description: Share token redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareTokenRedeemResponse"
  /v1/lists/{shareToken}/share-tokens:
    parameters:
      - $ref: "#/components/parameters/ShareToken"
      - $ref: "#/components/parameters/DeviceIdHeader"
    post:
      summary: Create additional share token for a list
      responses:
        "201":
          description: Share token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareTokenCreateResponse"